+++ src/endpoint/retrieve.js
@@ -5,8 +5,11 @@ const Constants = require('../constants');
 const Utils = require('../utils');
 
 
-async function retrieveRecords(from, to) {
+async function retrieveRecords(from, to, retrieveTimezone = false) {
   this.debug('retrieving records');
+
+  if (typeof retrieveTimezone !== 'boolean')
+    throw new TypeError(`The "retrieveTimezone" argument must be a "boolean". Received "${typeof retrieveTimezone}".`);
 
   const client = this.kit.client;
   const generated = this.generated;
@@ -30,8 +33,11 @@ async function retrieveRecords(from, to) {
         Object.assign(operation, operation.sample);
         operation[DATE] = DateTime.fromMillis(operation[TIMESTAMP] * 1000).toJSDate();
         generated.forEach((key) => delete operation[key]);
+
+        if (retrieveTimezone === true) operation[TIMEZONE] = `utc${operation[TIMEZONE]}`;
+        else delete operation[TIMEZONE];
+
         delete operation[TIMESTAMP];
-        delete operation[TIMEZONE];
         delete operation.sample;
 
         return operation;