+++ test/helpers/index.js
@@ -79,7 +79,10 @@ class Stream extends stream.Readable {
 
 function registerEndpoint() {
   if (!this.seed) {
-    const suite = new Error().stack.match(/\/(craft-ai-kit-energy\/.*)\.test\.js/)[1];
+    const suite = new Error().stack
+      // Replace Windows directory separator with Linux one preventing snapshot mismatch.
+      .replace(/\\/g, '/')
+      .match(/\/(craft-ai-kit-energy\/.*)\.test\.js/)[1];
 
     this.seed = `${suite}\n${this.id}`;
   }
@@ -93,9 +96,12 @@ function registerEndpoint() {
 }
 
 
+const IANA_ZONES = ['America/Adak', 'America/Chihuahua', 'America/St_Johns', 'Europe/Paris', 'Asia/Kolkata', 'Antarctica/South_Pole'];
 const INVALID_ARRAYS = [null, 0, true, 'string', Symbol(), new Uint8Array(10), () => {}];
+const INVALID_BOOLEANS = ['true', [1], [0], 'false', {}, new Date(), null];
 const INVALID_DATES = [false, NaN, 'N/A', 'NaN', 'unknown', '123456', 'string', '5151-51-51T51:51:51.515Z', () => {}];
 const INVALID_FUNCTIONS = [null, {}, 'abc', Boolean(), Array.from(2), new Date(2018)];
+const INVALID_IANA_ZONES = [null, false, 2, 'Europe/Taipei', 'America/Madrid'];
 const INVALID_NUMBERS = [null, {}, '12', true, new Date(2018), () => {}];
 const INVALID_OBJECTS = [null, 0, true, 'string', Symbol(), () => {}];
 const INVALID_STRINGS = [null, {}, [false], true, new Date(2018), Promise.resolve('string'), () => {}];
@@ -112,9 +118,12 @@ module.exports = {
   destroyProviderContext,
   identity,
   streamify,
+  IANA_ZONES,
   INVALID_ARRAYS,
+  INVALID_BOOLEANS,
   INVALID_DATES,
   INVALID_FUNCTIONS,
+  INVALID_IANA_ZONES,
   INVALID_NUMBERS,
   INVALID_OBJECTS,
   INVALID_STRINGS,