+++ test/kit.test.js
@@ -1,5 +1,6 @@
 const test = require('ava');
 
+const EnergyKit = require('../src/index');
 const Helpers = require('./helpers');
 
 
@@ -28,6 +29,7 @@ test('fails loading an endpoint with invalid definition', (t) => {
     .concat(INVALID_ORIGINS.map((value) => ({ id: 'id', energy: { origin: value } })))
     .concat(VALID_ORIGINS.map((value) => ({ id: 'id', energy: { origin: value } })))
     .concat(INVALID_OBJECTS.map((value) => ({ id: 'id', metadata: value })))
+    .concat(INVALID_IANA_ZONES.map((value) => ({ id: 'id', metadata: { zone: value } })))
     .map((definition) => t.throwsAsync(kit.loadEndpoint(definition))));
 });
 
@@ -66,6 +68,33 @@ test('loads an endpoint', async(t) => {
   t.notThrowsAsync(agentB);
   t.deepEqual(await agentA, await agentB);
   t.snapshot(endpointA);
+});
+
+test('initializes zone-less endpoint with the zone of the kit', (t) => {
+  const METADATA_OBJECTS = [{}, undefined];
+
+  return EnergyKit
+    .initialize({ zone: IANA_ZONES[0] })
+    .then((kit) => Promise
+      .all(METADATA_OBJECTS.map((metadata) => kit.loadEndpoint({ metadata, id: t.context.endpoint.register() })))
+      .then((endpoints) => endpoints.forEach((endpoint) => {
+        t.truthy(endpoint.metadata);
+        t.is(typeof endpoint.metadata, 'object');
+        t.is(endpoint.metadata.zone, kit.configuration.zone);
+      })));
+});
+
+test('initializes the endpoint with a valid zone', async(t) => {
+  const zone = IANA_ZONES[0];
+
+  return EnergyKit
+    .initialize({ zone: IANA_ZONES[IANA_ZONES.length - 1] })
+    .then((kit) => kit.loadEndpoint({ metadata: { zone }, id: t.context.endpoint.register() }))
+    .then((endpoint) => {
+      t.truthy(endpoint.metadata);
+      t.is(typeof endpoint.metadata, 'object');
+      t.is(endpoint.metadata.zone, zone);
+    });
 });
 
 test('loads an endpoint by forcing the recreation of the agent', async(t) => {
@@ -147,6 +176,8 @@ test('closes the kit', (t) => {
 });
 
 
+const IANA_ZONES = Helpers.IANA_ZONES;
+const INVALID_IANA_ZONES = Helpers.INVALID_IANA_ZONES;
 const INVALID_NUMBERS = Helpers.INVALID_NUMBERS;
 const INVALID_OBJECTS = Helpers.INVALID_OBJECTS;
 const INVALID_STRINGS = Helpers.INVALID_STRINGS;